pyproject.toml
# [tool.pyright]
# executionEnvironments = [{ root = "back" }]

[tool.basedpyright]
typeCheckingMode = "basic"
reportIncompatibleMethodOverride = false
reportIncompatibleVariableOverride = false
reportArgumentType = false
# reportUnreachable = true
# reportCallIssue = false
executionEnvironments = [{ root = "back" }]

[tool.pytest.ini_options]
addopts = "--strict-markers"
DJANGO_SETTINGS_MODULE = "back.settings"
python_files = ["test_*.py"]
markers = ["slow: marks tests as slow (deselect with '-m \"not slow\"')"]

[tool.coverage.run]
omit = ["manage.py"]

[build-system]
requires = ["setuptools==65.5.0"]
build-backend = "setuptools.build_meta"

[project]
name = "aigar"
version = "24.4.2"
authors = [{ name = "iCarto", email = "info@icarto.es" }]
readme = "README.md"
classifiers = [
    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
    "Private :: Do Not Upload",
]
requires-python = "==3.11.*"
dependencies = [
    "Django==4.2",
    "django-cors-headers==3.14.0",
    "django-environ==0.10.0",
    "django-filter==23.1",
    "django-spa==0.3.6",
    "djangorestframework==3.14.0",
    "django-solo==2.1.0",
    "djangorestframework-simplejwt==5.2.2",
    "platformdirs==3.10.0",
    "python-dateutil==2.8.2",
    "whitenoise==5.3.0",
    "django-import-export==3.3.1",
]

# https://stackoverflow.com/questions/75490231/
[project.optional-dependencies]
desktop = ["django-debug-toolbar==4.0.0", "django-extensions==3.2.1"]
test = [
    "pytest==7.4.0",
    "pytest-django==4.5.2",
    "pytest-cov==4.1.0",
    "selenium==4.14.0",
    "webdriver-manager==4.0.1",
    "factory_boy==3.3.0",
]
dev = [
    "django-debug-toolbar==4.0.0",
    "django-extensions==3.2.1",
    "vulture==2.7",
    "django-stubs==4.2.5",
    "djangorestframework-stubs",
    # "aigar[test]",
]

all = ["aigar[dev, test]"]

[tool.setuptools]
package-dir = { "aigar" = "back" }
include-package-data = true

[tool.setuptools.package-data]
# *.css *.woff *.svg *.js *.gif *.ico *.png *.jpg *.html *.js.map
"aigar.static" = ["**"]
"aigar.templates" = ["**"]

[tool.vulture]
# Configuration for vulture: https://github.com/jendrikseipp/vulture
# Install in your virtual environment and run:
# python -m vulture | tail -r | less
# The below configuration tries to remove some false positives, but there are
# still many, for example for model properties used only in templates.
# See also:
# https://adamj.eu/tech/2023/07/12/django-clean-up-unused-code-vulture/
exclude = ["*/settings.py", "*/settings/*.py", "*/migrations/*.py"]
ignore_decorators = [
    # Django
    "@receiver",
    "@register.filter",
    "@register.inclusion_tag",
    "@register.simple_tag",
    # django.contrib.admin
    "@admin.action",
    "@admin.display",
    "@admin.register",
    # pytest
    "@pytest.fixture",
    "@factory.post_generation",
]
ignore_names = [
    # Django
    "*Config",     # AppConfig subclasses
    "*Middleware",
    "clean_*",
    "Meta",
    "urlpatterns",
    # django.contrib.admin
    "get_extra",
    "get_fieldsets",
    "has_add_permission",
    "has_change_permission",
    "has_delete_permission",
    "has_view_permission",
    "lookups",

]
paths = ["back"]
sort_by_size = true
verbose = true
min_confidence = 0


[tool.ruff]
extend = "./config/pyproject.toml"

extend-exclude = [
    "/**/migrations/*", # folders excluded when working with django
]

[tool.ruff.lint]

extend-ignore = ["ANN", "ERA", "TCH"]

[tool.ruff.lint.isort]
# https://docs.astral.sh/ruff/settings/#isort-known-third-party
known-first-party = [
    "users",
    "app",
    "back",
    "domains",
] # Ajustar en cada proyecto
